Reorder HadREM fixes
