+++ esmvalcore/cmor/_fixes/cordex/cnrm_cerfacs_cnrm_cm5/mohc_hadrem3_ga7_05.py
@@ -1,5 +1,5 @@
 """Fixes for rcm MOHC-HadREM3-GA7-05 driven by CNRM-CERFACS-CNRM-CM5."""
-from ..mohc_hadgem2_es.mohc_hadrem3_ga7_05 import Tas as BaseFix
+from ..cordex_fixes import MOHCHadREM3GA705 as BaseFix
 
 Tas = BaseFix
