+++ esmvalcore/cmor/_fixes/cordex/cordex_fixes.py
@@ -1,0 +1,28 @@
+"""Fixes that are shared between datasets and drivers."""
+from ..fix import Fix
+
+
+class MOHCHadREM3GA705(Fix):
+    """General fix for MOHC-HadREM3-GA7-05."""
+
+    def fix_metadata(self, cubes):
+        """
+        Fix time long_name.
+        Fix latitude and longitude var_name.
+
+        Parameters
+        ----------
+        cubes : iris.cube.CubeList
+            Input cubes.
+
+        Returns
+        -------
+        iris.cube.CubeList
+
+        """
+        for cube in cubes:
+            cube.coord('latitude').var_name = 'lat'
+            cube.coord('longitude').var_name = 'lon'
+            cube.coord('time').long_name = 'time'
+
+        return cubes
