+++ esmvalcore/cmor/_fixes/cordex/ncc_noresm1_m/mohc_hadrem3_ga7_05.py
@@ -1,5 +1,5 @@
 """Fixes for rcm HadREM3-GA7-05 driven by NCC-NorESM1-M."""
-from ..mohc_hadgem2_es.mohc_hadrem3_ga7_05 import Tas as BaseFix
+from ..cordex_fixes import MOHCHadREM3GA705 as BaseFix
 
 Tas = BaseFix
